+++ node/internal/streams/duplex.mjs
@@ -2,16 +2,12 @@
 // Copyright Joyent and Node contributors. All rights reserved. MIT license.
 // deno-lint-ignore-file
 
-import { validateObject, validateBoolean } from "../validators.mjs";
 import { createDeferredPromise } from "../util.mjs";
 import { destroyer } from "./destroy.mjs";
 import { isBlob } from "../blob.mjs";
-import { Buffer } from "../../buffer.ts";
 import {
   AbortError,
   ERR_INVALID_ARG_TYPE,
-  ERR_INVALID_ARG_VALUE,
-  ERR_STREAM_PREMATURE_CLOSE,
   ERR_INVALID_RETURN_VALUE,
 } from "../errors.ts";
 import {
@@ -19,10 +15,8 @@ import {
   isIterable,
   isNodeStream,
   isReadable,
-  isReadableEnded,
   isReadableNodeStream,
   isWritable,
-  isWritableEnded,
   isWritableNodeStream,
 } from "./utils.mjs";
 import * as process from "../../_process/process.ts";
@@ -30,6 +24,7 @@ import _from from "./from.mjs";
 import eos from "./end-of-stream.mjs";
 import Readable from "./readable.mjs";
 import Writable from "./writable.mjs";
+import { newStreamDuplexFromReadableWritablePair } from "../webstreams/adapters.mjs";
 
 Object.setPrototypeOf(Duplex.prototype, Readable.prototype);
 Object.setPrototypeOf(Duplex, Readable);
@@ -123,193 +118,7 @@ Object.defineProperties(Duplex.prototype, {
   },
 });
 
-Duplex.fromWeb = function (pair, options) {
-  validateObject(pair, 'pair');
-  const {
-    readable: readableStream,
-    writable: writableStream,
-  } = pair;
-
-  if (!isReadableStream(readableStream)) {
-    throw new ERR_INVALID_ARG_TYPE(
-      'pair.readable',
-      'ReadableStream',
-      readableStream);
-  }
-  if (!isWritableStream(writableStream)) {
-    throw new ERR_INVALID_ARG_TYPE(
-      'pair.writable',
-      'WritableStream',
-      writableStream);
-  }
-
-  validateObject(options, 'options');
-  const {
-    allowHalfOpen = false,
-    objectMode = false,
-    encoding,
-    decodeStrings = true,
-    highWaterMark,
-    signal,
-  } = options;
-
-  validateBoolean(objectMode, 'options.objectMode');
-  if (encoding !== undefined && !Buffer.isEncoding(encoding))
-    throw new ERR_INVALID_ARG_VALUE(encoding, 'options.encoding');
-
-  const writer = writableStream.getWriter();
-  const reader = readableStream.getReader();
-  let writableClosed = false;
-  let readableClosed = false;
-
-  const duplex = new Duplex({
-    allowHalfOpen,
-    highWaterMark,
-    objectMode,
-    encoding,
-    decodeStrings,
-    signal,
-
-    writev(chunks, callback) {
-      function done(error) {
-        error = error.filter((e) => e);
-        try {
-          callback(error.length === 0 ? undefined : error);
-        } catch (error) {
-          // In a next tick because this is happening within
-          // a promise context, and if there are any errors
-          // thrown we don't want those to cause an unhandled
-          // rejection. Let's just escape the promise and
-          // handle it separately.
-          process.nextTick(() => destroy(duplex, error));
-        }
-      }
-
-      writer.ready.then(
-        () => Promise.All(
-          chunks.map((data) => writer.write(data.chunk))
-        ).then(done, done),
-        done,
-      );
-    },
-
-    write(chunk, encoding, callback) {
-      if (typeof chunk === 'string' && decodeStrings && !objectMode) {
-        chunk = Buffer.from(chunk, encoding);
-        chunk = new Uint8Array(
-          chunk.buffer,
-          chunk.byteOffset,
-          chunk.byteLength);
-      }
-
-      function done(error) {
-        try {
-          callback(error);
-        } catch (error) {
-          destroy(duplex, error);
-        }
-      }
-
-      writer.ready.then(
-        () => writer.write(chunk).then(done, done),
-        done,
-      );
-    },
-
-    final(callback) {
-      function done(error) {
-        try {
-          callback(error);
-        } catch (error) {
-          // In a next tick because this is happening within
-          // a promise context, and if there are any errors
-          // thrown we don't want those to cause an unhandled
-          // rejection. Let's just escape the promise and
-          // handle it separately.
-          process.nextTick(() => destroy(duplex, error));
-        }
-      }
-
-      if (!writableClosed) {
-        writer.close().then(done, done);
-      }
-    },
-
-    read() {
-      reader.read().then(
-        (chunk) => {
-          if (chunk.done) {
-            duplex.push(null);
-          } else {
-            duplex.push(chunk.value);
-          }
-        },
-        (error) => destroy(duplex, error));
-    },
-
-    destroy(error, callback) {
-      function done() {
-        try {
-          callback(error);
-        } catch (error) {
-          // In a next tick because this is happening within
-          // a promise context, and if there are any errors
-          // thrown we don't want those to cause an unhandled
-          // rejection. Let's just escape the promise and
-          // handle it separately.
-          process.nextTick(() => { throw error; });
-        }
-      }
-
-      async function closeWriter() {
-        if (!writableClosed)
-          await writer.abort(error);
-      }
-
-      async function closeReader() {
-        if (!readableClosed)
-          await reader.cancel(error);
-      }
-
-      if (!writableClosed || !readableClosed) {
-        Promise.All([
-          closeWriter(),
-          closeReader(),
-        ]).then(done, done);
-        return;
-      }
-
-      done();
-    },
-  });
-
-  writer.closed.then(
-    () => {
-      writableClosed = true;
-      if (!isWritableEnded(duplex))
-        destroy(duplex, new ERR_STREAM_PREMATURE_CLOSE());
-    },
-    (error) => {
-      writableClosed = true;
-      readableClosed = true;
-      destroy(duplex, error);
-    });
-
-  reader.closed.then(
-    () => {
-      readableClosed = true;
-      if (!isReadableEnded(duplex))
-        duplex.push(null);
-    },
-    (error) => {
-      writableClosed = true;
-      readableClosed = true;
-      destroy(duplex, error);
-    });
-
-  return duplex;
-}
-
+Duplex.fromWeb = newStreamDuplexFromReadableWritablePair;
 ///////////////////////////
 // This is part of the duplexify module, but had to be brought into scope
 // to avoid a circular dependency
