+++ node/internal/webstreams/adapters.mjs
@@ -1,0 +1,219 @@
+import { validateBoolean, validateObject } from "../validators.mjs";
+import {
+  ERR_INVALID_ARG_TYPE,
+  ERR_INVALID_ARG_VALUE,
+  ERR_STREAM_PREMATURE_CLOSE,
+} from "../errors.ts";
+import { Buffer } from "../../buffer.ts";
+import * as process from "../../_process/process.ts";
+import { isReadableEnded, isWritableEnded } from "../streams/utils.mjs";
+import Duplex from "../streams/duplex.mjs";
+
+function isReadableStream(object) {
+  return object instanceof ReadableStream;
+}
+
+function isWritableStream(object) {
+  return object instanceof WritableStream;
+}
+
+export function newStreamDuplexFromReadableWritablePair(pair, options) {
+  validateObject(pair, "pair");
+  const {
+    readable: readableStream,
+    writable: writableStream,
+  } = pair;
+
+  if (!isReadableStream(readableStream)) {
+    throw new ERR_INVALID_ARG_TYPE(
+      "pair.readable",
+      "ReadableStream",
+      readableStream,
+    );
+  }
+  if (!isWritableStream(writableStream)) {
+    throw new ERR_INVALID_ARG_TYPE(
+      "pair.writable",
+      "WritableStream",
+      writableStream,
+    );
+  }
+
+  validateObject(options, "options");
+  const {
+    allowHalfOpen = false,
+    objectMode = false,
+    encoding,
+    decodeStrings = true,
+    highWaterMark,
+    signal,
+  } = options;
+
+  validateBoolean(objectMode, "options.objectMode");
+  if (encoding !== undefined && !Buffer.isEncoding(encoding)) {
+    throw new ERR_INVALID_ARG_VALUE(encoding, "options.encoding");
+  }
+
+  const writer = writableStream.getWriter();
+  const reader = readableStream.getReader();
+  let writableClosed = false;
+  let readableClosed = false;
+
+  const duplex = new Duplex({
+    allowHalfOpen,
+    highWaterMark,
+    objectMode,
+    encoding,
+    decodeStrings,
+    signal,
+
+    writev(chunks, callback) {
+      function done(error) {
+        error = error.filter((e) => e);
+        try {
+          callback(error.length === 0 ? undefined : error);
+        } catch (error) {
+          // In a next tick because this is happening within
+          // a promise context, and if there are any errors
+          // thrown we don't want those to cause an unhandled
+          // rejection. Let's just escape the promise and
+          // handle it separately.
+          process.nextTick(() => destroy(duplex, error));
+        }
+      }
+
+      writer.ready.then(
+        () =>
+          Promise.All(
+            chunks.map((data) => writer.write(data.chunk)),
+          ).then(done, done),
+        done,
+      );
+    },
+
+    write(chunk, encoding, callback) {
+      if (typeof chunk === "string" && decodeStrings && !objectMode) {
+        chunk = Buffer.from(chunk, encoding);
+        chunk = new Uint8Array(
+          chunk.buffer,
+          chunk.byteOffset,
+          chunk.byteLength,
+        );
+      }
+
+      function done(error) {
+        try {
+          callback(error);
+        } catch (error) {
+          destroy(duplex, error);
+        }
+      }
+
+      writer.ready.then(
+        () => writer.write(chunk).then(done, done),
+        done,
+      );
+    },
+
+    final(callback) {
+      function done(error) {
+        try {
+          callback(error);
+        } catch (error) {
+          // In a next tick because this is happening within
+          // a promise context, and if there are any errors
+          // thrown we don't want those to cause an unhandled
+          // rejection. Let's just escape the promise and
+          // handle it separately.
+          process.nextTick(() => destroy(duplex, error));
+        }
+      }
+
+      if (!writableClosed) {
+        writer.close().then(done, done);
+      }
+    },
+
+    read() {
+      reader.read().then(
+        (chunk) => {
+          if (chunk.done) {
+            duplex.push(null);
+          } else {
+            duplex.push(chunk.value);
+          }
+        },
+        (error) => destroy(duplex, error),
+      );
+    },
+
+    destroy(error, callback) {
+      function done() {
+        try {
+          callback(error);
+        } catch (error) {
+          // In a next tick because this is happening within
+          // a promise context, and if there are any errors
+          // thrown we don't want those to cause an unhandled
+          // rejection. Let's just escape the promise and
+          // handle it separately.
+          process.nextTick(() => {
+            throw error;
+          });
+        }
+      }
+
+      async function closeWriter() {
+        if (!writableClosed) {
+          await writer.abort(error);
+        }
+      }
+
+      async function closeReader() {
+        if (!readableClosed) {
+          await reader.cancel(error);
+        }
+      }
+
+      if (!writableClosed || !readableClosed) {
+        Promise.All([
+          closeWriter(),
+          closeReader(),
+        ]).then(done, done);
+        return;
+      }
+
+      done();
+    },
+  });
+
+  writer.closed.then(
+    () => {
+      writableClosed = true;
+      if (!isWritableEnded(duplex)) {
+        destroy(duplex, new ERR_STREAM_PREMATURE_CLOSE());
+      }
+    },
+    (error) => {
+      writableClosed = true;
+      readableClosed = true;
+      destroy(duplex, error);
+    },
+  );
+
+  reader.closed.then(
+    () => {
+      readableClosed = true;
+      if (!isReadableEnded(duplex)) {
+        duplex.push(null);
+      }
+    },
+    (error) => {
+      writableClosed = true;
+      readableClosed = true;
+      destroy(duplex, error);
+    },
+  );
+
+  return duplex;
+}
